+++ kaithem/src/builtintags.py
@@ -8,7 +8,7 @@ def civilTwilight():
             return 1
         else:
             return 0
-    except:
+    except Exception:
         return -1
 
 
@@ -17,6 +17,7 @@ twilightTag.min = -1
 twilightTag.max = 1
 twilightTag.interval = 60
 twilightTag.description = "Unless overridden, 1 if dark, else 0, -1 if no location is set"
+twilightTag.value = civilTwilight
 
 
 def night():
@@ -25,7 +26,7 @@ def night():
             return 1
         else:
             return 0
-    except:
+    except Exception:
         return -1
 
 
@@ -34,3 +35,21 @@ nTag.min = -1
 nTag.max = 1
 nTag.interval = 60
 nTag.description = "Unless overridden, 1 if night, else 0, -1 if no location is set"
+nTag.value = night
+
+
+def publicIP():
+    try:
+        import requests
+        r = requests.get("http://api.ipify.org/", timeout=1)
+        r.raise_for_status()
+        return r.text
+    except Exception:
+        print(traceback.format_exc())
+        return ""
+
+
+ipTag = tagpoints.StringTag("/system/network/publicIP")
+ipTag.interval = 3600
+ipTag.description = "The current public IP address, as seen by http://api.ipify.org.  If the server is unreachable, will be the empty string."
+ipTag.value = publicIP